+++ docs/changes.md
@@ -88,7 +88,8 @@ seconds
 ### Global
 * Player healthbars are disabled.
 
-### Bug Fixes
+### Bug Fixes & Improvements
 * ARCs now take corrode damage after deploying and undeploying.
 * Potential fix for view angles being set incorrectly on spawn 
 * Cysts should no longer starve when they shouldn't
+* Cysts should be able to find their parents more reliably when the chain is updated